requirements.txt
discord.py
python-dotenv
asyncpg
asyncio